+++ src/App.tsx
@@ -11,7 +11,7 @@ import { PatientAllergies } from "@/components/content/allergies/patient-allergi
 import { PatientImmunizations } from "./components/content/immunizations/patient-immunizations";
 import { PatientCareTeam } from "./components/content/care-team/patient-careteam";
 import { PatientDocuments } from "./components/content/document/patient-documents";
-import { PatientsTable } from "./components/content/patients/patients-table";
+import { ZusAggregatedProfile } from "@/components/content/zus-aggregated-profile/zus-aggregated-profile";
 
 const {
   VITE_SYSTEM_URL,
@@ -43,17 +43,32 @@ type DemoComponent = {
 const demoComponents: DemoComponent[] = [
   { render: () => <PatientCareTeam />, title: "Patient CareTeam" },
   {
-    render: () => <PatientMedications handleAddToRecord={() => null} />,
-    title: "Patient Medications Tabbed",
+    render: () => (
+      <ZusAggregatedProfile
+        resources={[
+          "medications",
+          "timelines",
+          "allergies",
+          "conditions-outside",
+        ]}
+      />
+    ),
+    title: "ZAP",
   },
-  { render: () => <PatientConditions />, title: "Patient Conditions 2.0" },
+  {
+    render: () => <PatientMedications handleAddToRecord={() => null} />,
+    title: "Patient Medications",
+  },
+  {
+    render: () => <PatientConditions />,
+    title: "Patient Conditions 2.0",
+  },
   {
     render: () => <PatientDocuments />,
     title: "Patient Documents",
   },
   { render: () => <PatientAllergies />, title: "Patient Allergies" },
   { render: () => <PatientImmunizations />, title: "Patient Immunizations" },
-  { render: () => <PatientsTable />, title: "Patient Table" },
 ];
 
 const DemoApp = ({ accessToken = "" }) => (