+++ src/components/content/conditions/patient-conditions.tsx
@@ -22,13 +22,13 @@ import {
 } from "@/fhir/conditions";
 import { useBreakpoints } from "@/hooks/use-breakpoints";
 import "./patient-conditions.scss";
-import { applyFilters } from "@/utils/filters";
 
 export type PatientConditionsProps = {
   className?: string;
   readOnly?: boolean;
   hideBuilderOwnedRecords?: boolean;
   hideOutsideOwnedRecords?: boolean;
+  hideRequestRecords?: boolean;
 };
 
 export const PatientConditions = withErrorBoundary(
@@ -37,12 +37,14 @@ export const PatientConditions = withErrorBoundary(
     readOnly = false,
     hideBuilderOwnedRecords = false,
     hideOutsideOwnedRecords = false,
+    hideRequestRecords = false,
   }: PatientConditionsProps) => {
     // State.
     const [collection, setCollection] = useState<FilterCollection>(
       hideBuilderOwnedRecords ? "other" : "patient"
     );
-    const { filters, updateFilters } = useConditionFilters(collection);
+    const { filters, updateFilters, applyFilters } =
+      useConditionFilters(collection);
     const { applySorts, sortOptions, updateSorts, currentSorts } =
       useConditionSorts(collection);
     const containerRef = useRef<HTMLDivElement>(null);
@@ -65,11 +67,9 @@ export const PatientConditions = withErrorBoundary(
     const patientConditions = patientConditionsQuery.data ?? [];
     const otherConditions = otherConditionsQuery.data;
 
-    let conditions = applyFilters(
-      collection === "patient" ? patientConditions : otherConditions,
-      filters[collection]
+    const conditions = applySorts(
+      applyFilters(patientConditions, otherConditions)
     );
-    conditions = applySorts(conditions);
     const RowActions =
       collection === "patient"
         ? PatientConditionHoverActions
@@ -101,6 +101,7 @@ export const PatientConditions = withErrorBoundary(
           )}
           setFilters={updateFilters}
           filters={filters[collection]}
+          hideRequestRecords={hideRequestRecords || collection === "patient"}
         />
 
         <div className="ctw-overflow-hidden">