+++ src/components/content/medications/provider-meds-table.tsx
@@ -4,10 +4,14 @@ import {
   MedicationsTableBase,
   MedsHistoryTempProps,
 } from "@/components/content/medications-table-base";
-import { useMedicationSorts } from "@/components/content/medications/patient-medications-sort";
+import {
+  defaultMedicationSort,
+  medicationSortOptions,
+} from "@/components/content/medications/patient-medications-sort";
 import { withErrorBoundary } from "@/components/core/error-boundary";
 import { SortButton } from "@/components/core/sort-button/sort-button";
 import { MedicationStatementModel } from "@/fhir/models/medication-statement";
+import { useFilteredSortedData } from "@/hooks/use-filtered-sorted-data";
 import { useQueryAllPatientMedications } from "@/hooks/use-medications";
 import { get, isFunction, pipe, toLower } from "@/utils/nodash/fp";
 import { sort, SortDir } from "@/utils/sort";
@@ -40,8 +44,11 @@ export const ProviderMedsTable = withErrorBoundary(
     >([]);
     const { builderMedications, isLoading } = useQueryAllPatientMedications();
     const openMedHistoryDrawer = useMedicationHistory();
-    const { currentSorts, updateSorts, sortOptions, applySorts } =
-      useMedicationSorts();
+    const { data, setSort } = useFilteredSortedData({
+      defaultFilters: {},
+      defaultSort: defaultMedicationSort,
+      records: builderMedications,
+    });
 
     function openHistoryDrawer(row: MedicationStatementModel) {
       // Temp - onOpen and onAfterOpen should be side-effect free as
@@ -59,31 +66,30 @@ export const ProviderMedsTable = withErrorBoundary(
     }
 
     useEffect(() => {
-      if (!builderMedications) return;
       setMedicationModels(
         sort(
           showInactive
-            ? builderMedications
-            : builderMedications.filter((bm) => bm.displayStatus === "Active"),
+            ? data
+            : data.filter((bm) => bm.displayStatus === "Active"),
           pipe(get(sortColumn), toLower),
           sortOrder
         )
       );
-    }, [builderMedications, sortColumn, sortOrder, showInactive]);
+    }, [data, sortColumn, sortOrder, showInactive]);
 
     return (
       <>
         <div className="ctw-flex ctw-flex-wrap ctw-gap-x-2">
           <SortButton
             className="ctw-my-2"
-            options={sortOptions}
-            updateSorts={updateSorts}
-            currentSorts={currentSorts}
+            options={medicationSortOptions}
+            onChange={setSort}
+            defaultSort={defaultMedicationSort}
           />
         </div>
 
         <MedicationsTableBase
-          medicationStatements={applySorts(medicationModels)}
+          medicationStatements={medicationModels}
           telemetryNamespace="ProviderMedsTable"
           isLoading={isLoading}
           handleRowClick={openHistoryDrawer}